Made the second task a bit more accurate with long double but precision is not yet perfect

+++ A/21_Nikolay_Danailov/homework_3/task_2.c
@@ -2,9 +2,9 @@
 
 int main()
 {
-	double firstNumber = 2;
-	double secondNumber = 1;
-	double thirdNumber;
+	long double firstNumber = 2;
+	long double secondNumber = 1;
+	long double thirdNumber;
 	
 	int i;
 	for (i = 2; i <= 100; i += 1)
@@ -12,7 +12,7 @@ int main()
 		thirdNumber = secondNumber + firstNumber;
 		firstNumber = secondNumber;
 		secondNumber = thirdNumber;
-		printf("%.0lf, ", thirdNumber);
+		printf("%Lf, ", thirdNumber);
 	}
 	
 	return 0;
